+++ lib/main.dart
@@ -11,9 +11,9 @@ class AppWidget extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return const MaterialApp(
       title: 'Randomizer',
-      home: const RangeSelectorPage(),
+      home: RangeSelectorPage(),
     );
   }
 }
