+++ lib/range_selector_form.dart
@@ -1,0 +1,74 @@
+import 'package:flutter/material.dart';
+
+typedef IntValueSetter = void Function(int value);
+
+class RangeSelectorForm extends StatelessWidget {
+  const RangeSelectorForm({
+    Key? key,
+    required this.formKey,
+    required this.minValueSetter,
+    required this.maxValueSetter,
+  }) : super(key: key);
+
+  final GlobalKey<FormState> formKey;
+  final IntValueSetter minValueSetter;
+  final IntValueSetter maxValueSetter;
+
+  @override
+  Widget build(BuildContext context) {
+    return Form(
+      key: formKey,
+      child: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            RangeSelectorTextFormField(
+              labelText: 'Minimum',
+              intValueSetter: minValueSetter,
+            ),
+            const SizedBox(
+              height: 12,
+            ),
+            RangeSelectorTextFormField(
+              labelText: 'Maximum',
+              intValueSetter: maxValueSetter,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class RangeSelectorTextFormField extends StatelessWidget {
+  const RangeSelectorTextFormField({
+    required this.labelText,
+    required this.intValueSetter,
+    Key? key,
+  }) : super(key: key);
+  final String labelText;
+  final IntValueSetter intValueSetter;
+
+  @override
+  Widget build(BuildContext context) {
+    return TextFormField(
+      decoration: InputDecoration(
+        border: const OutlineInputBorder(),
+        labelText: labelText,
+      ),
+      keyboardType: const TextInputType.numberWithOptions(
+        decimal: false,
+        signed: true,
+      ),
+      validator: (value) {
+        if (value == null || int.tryParse(value) == null) {
+          return 'Must be an integer';
+        } else {
+          return null;
+        }
+      },
+      onSaved: (newValue) => intValueSetter(int.parse(newValue ?? '')),
+    );
+  }
+}
