+++ lib/range_selector_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_foundations/range_selector_form.dart';
 
 class RangeSelectorPage extends StatefulWidget {
   const RangeSelectorPage({super.key});
@@ -18,27 +19,10 @@ class _RangeSelectorPageState extends State<RangeSelectorPage> {
       appBar: AppBar(
         title: const Text('Select Range'),
       ),
-      body: Form(
-        key: formKey,
-        child: Padding(
-          padding: const EdgeInsets.all(16.0),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              RangeSelectorTextFormField(
-                labelText: 'Minimum',
-                intValueSetter: (value) => _min = value,
-              ),
-              const SizedBox(
-                height: 12,
-              ),
-              RangeSelectorTextFormField(
-                labelText: 'Maximum',
-                intValueSetter: (value) => _max = value,
-              ),
-            ],
-          ),
-        ),
+      body: RangeSelectorForm(
+        formKey: formKey,
+        minValueSetter: (value) => _min = value,
+        maxValueSetter: (value) => _max = value,
       ),
       floatingActionButton: FloatingActionButton(
         child: const Icon(Icons.arrow_forward),
@@ -49,38 +33,6 @@ class _RangeSelectorPageState extends State<RangeSelectorPage> {
           // TODO: Navigate to the generator page
         },
       ),
-    );
-  }
-}
-
-class RangeSelectorTextFormField extends StatelessWidget {
-  const RangeSelectorTextFormField({
-    required this.labelText,
-    required this.intValueSetter,
-    Key? key,
-  }) : super(key: key);
-  final String labelText;
-  final void Function(int value) intValueSetter;
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      decoration: InputDecoration(
-        border: const OutlineInputBorder(),
-        labelText: labelText,
-      ),
-      keyboardType: const TextInputType.numberWithOptions(
-        decimal: false,
-        signed: true,
-      ),
-      validator: (value) {
-        if (value == null || int.tryParse(value) == null) {
-          return 'Must be an integer';
-        } else {
-          return null;
-        }
-      },
-      onSaved: (newValue) => intValueSetter(int.parse(newValue ?? '')),
     );
   }
 }
